Converted Element types to layouts like those in Android Settings and to take layouts as input with specific IDs (see samples). Added DialogActivity class derived from Activity to handle the adapter and Root elements. Added a few elements, fixed elements so that scrolling and rotation work (in the DialogActivity class anyway, values lost on rotation in Activity1.

+++ MonoDroid.Dialog/Reflect.cs
@@ -126,11 +126,11 @@ namespace MonoDroid.Dialog
 
     public class RangeAttribute : Attribute
     {
-        public float High;
-        public float Low;
+        public int High;
+        public int Low;
         public bool ShowCaption;
 
-        public RangeAttribute(float low, float high)
+        public RangeAttribute(int low, int high)
         {
             Low = low;
             High = high;
